Created utils and adding contact info
