+++ lib/utils.dart
@@ -1,0 +1,131 @@
+import 'dart:math';
+
+import 'package:flutter/services.dart';
+
+class DateTextFormatter extends TextInputFormatter {
+  static const _maxChars = 8;
+
+  @override
+  TextEditingValue formatEditUpdate(
+    TextEditingValue oldValue,
+    TextEditingValue newValue,
+  ) {
+    String separator = '/';
+    var text = _format(
+      newValue.text,
+      oldValue.text,
+      separator,
+    );
+
+    return newValue.copyWith(
+      text: text,
+      selection: updateCursorPosition(
+        oldValue,
+        text,
+      ),
+    );
+  }
+
+  String _format(
+    String value,
+    String oldValue,
+    String separator,
+  ) {
+    var isErasing = value.length < oldValue.length;
+    var isComplete = value.length > _maxChars + 2;
+
+    if (!isErasing && isComplete) {
+      return oldValue;
+    }
+
+    value = value.replaceAll(separator, '');
+    final result = <String>[];
+
+    for (int i = 0; i < min(value.length, _maxChars); i++) {
+      result.add(value[i]);
+      if ((i == 1 || i == 3) && i != value.length - 1) {
+        result.add(separator);
+      }
+    }
+
+    return result.join();
+  }
+
+  TextSelection updateCursorPosition(
+    TextEditingValue oldValue,
+    String text,
+  ) {
+    var endOffset = max(
+      oldValue.text.length - oldValue.selection.end,
+      0,
+    );
+
+    var selectionEnd = text.length - endOffset;
+
+    return TextSelection.fromPosition(TextPosition(offset: selectionEnd));
+  }
+}
+
+class PhoneNumberFormatter extends TextInputFormatter {
+  static const _maxChars = 10;
+
+  @override
+  TextEditingValue formatEditUpdate(
+    TextEditingValue oldValue,
+    TextEditingValue newValue,
+  ) {
+    String separator = '-';
+    var text = _format(
+      newValue.text,
+      oldValue.text,
+      separator,
+    );
+
+    return newValue.copyWith(
+      text: text,
+      selection: updateCursorPosition(
+        oldValue,
+        text,
+      ),
+    );
+  }
+
+  String _format(
+    String value,
+    String oldValue,
+    String separator,
+  ) {
+    var isErasing = value.length < oldValue.length;
+    var isComplete = value.length > _maxChars + 2;
+
+    if (!isErasing && isComplete) {
+      return oldValue;
+    }
+
+    value = value.replaceAll(separator, '');
+    final result = <String>[];
+
+    for (int i = 0; i < min(value.length, _maxChars); i++) {
+      result.add(value[i]);
+      if ((i == 2 || i == 5) && i != value.length - 1) {
+        result.add(separator);
+      }
+    }
+
+    return result.join();
+  }
+
+  TextSelection updateCursorPosition(
+    TextEditingValue oldValue,
+    String text,
+  ) {
+    var endOffset = max(
+      oldValue.text.length - oldValue.selection.end,
+      0,
+    );
+
+    var selectionEnd = text.length - endOffset;
+
+    return TextSelection.fromPosition(TextPosition(offset: selectionEnd));
+  }
+}
