+++ lib/views/patient/patient_login.dart
@@ -51,10 +51,7 @@ class Logo extends StatelessWidget {
             textAlign: TextAlign.center,
             style: isSmallScreen
                 ? Theme.of(context).textTheme.headline5
-                : Theme.of(context)
-                    .textTheme
-                    .headline4
-                    ?.copyWith(color: Colors.black),
+                : Theme.of(context).textTheme.headline4?.copyWith(color: Colors.black),
           ),
         ),
         gap(),
@@ -92,9 +89,8 @@ class FormContentState extends State<FormContent> {
                   return 'Please enter some text';
                 }
 
-                bool emailValid = RegExp(
-                        r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+")
-                    .hasMatch(value);
+                bool emailValid =
+                    RegExp(r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+").hasMatch(value);
                 if (!emailValid) {
                   return 'Please enter a valid email';
                 }
@@ -127,9 +123,7 @@ class FormContentState extends State<FormContent> {
                   prefixIcon: const Icon(Icons.lock_outline_rounded),
                   border: const OutlineInputBorder(),
                   suffixIcon: IconButton(
-                    icon: Icon(isPasswordVisible
-                        ? Icons.visibility_off
-                        : Icons.visibility),
+                    icon: Icon(isPasswordVisible ? Icons.visibility_off : Icons.visibility),
                     onPressed: () {
                       setState(() {
                         isPasswordVisible = !isPasswordVisible;
@@ -142,8 +136,7 @@ class FormContentState extends State<FormContent> {
               width: double.infinity,
               child: ElevatedButton(
                 style: ElevatedButton.styleFrom(
-                  shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(4)),
+                  shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(4)),
                 ),
                 child: const Padding(
                   padding: EdgeInsets.all(10.0),
@@ -154,9 +147,6 @@ class FormContentState extends State<FormContent> {
                 ),
                 onPressed: () {
                   Navigator.pushNamed(context, "/patient/home");
-                  // if (_formKey.currentState?.validate() ?? false) {
-                  //   /// do something
-                  // }
                 },
               ),
             ),
@@ -165,8 +155,7 @@ class FormContentState extends State<FormContent> {
               width: double.infinity,
               child: ElevatedButton(
                 style: ElevatedButton.styleFrom(
-                  shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(4)),
+                  shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(4)),
                 ),
                 child: const Padding(
                   padding: EdgeInsets.all(10.0),
